Client: implemented identity-aware-broadcast
Implemented command to send a broadcast message (procedure call) to an
identity-aware module (client-side and server-side).
On the server side, the GetIdentitySkeletonSetFunc is provided (on the previous
commit the GetIdentitySkeletonFunc had been provided).
On the server side, the same method is used as the one implemented for
the identity-aware-unicast.
On the client side, if an arc does not send the ACK it is reported.