Remove duty of choosing a linklocal address.
The task of choosing a new linklocal address for each new NIC (that
is a new pseudonic created in a new namespace when adding a new
identity to the system) does not fit in this module.
Furthermore, we feel now the need for changing the range which we
want to choose the new address inside.